refactor for local cosmosDB emulator

+++ OneStop/Model/IPEDS_CDEP.cs
@@ -8,7 +8,7 @@ namespace OneStopHelper.Model
         [JsonProperty(PropertyName = "id")]
         public string Id { get; set; }
         public string UNITID { get; set; }
-        public double year { get; set; }
+        public int year { get; set; }
         public Dictionary<string, IPEDS_CDEP_ITEM> Items { get; set; } //key is CIPCODE in CDEP table
 	}
 }
